Revert "feat(web): reach projects from a folded sidebar (#13943)"
This reverts commit d721efe3f980b1158fc61ad1b10afb4aade9848c.
