+++ web/lib/opal/src/components/buttons/line-item-button/components.tsx
@@ -14,7 +14,7 @@ import { type ContentActionProps, ContentAction } from "@opal/layouts";
 
 type ContentPassthroughProps = DistributiveOmit<
   ContentActionProps,
-  "width" | "ref"
+  "padding" | "width" | "ref"
 >;
 
 type LineItemButtonOwnProps = Pick<
@@ -93,7 +93,6 @@ function LineItemButton({
   // Sizing
   rounding = "md",
   width = "full",
-  padding = 0.5,
   tooltip,
   tooltipSide = "top",
 
@@ -129,11 +128,11 @@ function LineItemButton({
         rounding={rounding}
         {...rowButtonProps}
       >
-        <div className="w-full p-1.5">
+        <div className="w-full p-2">
           <ContentAction
             color="interactive"
             {...(contentActionProps as ContentActionProps)}
-            padding={padding}
+            padding={0}
           />
         </div>
       </Interactive.Container>
