+++ web/lib/opal/src/components/buttons/sidebar-tab/components.tsx
@@ -136,7 +136,7 @@ function SidebarTab({
     icon ??
     (nested
       ? ((() => (
-          <div className="w-4" aria-hidden="true" />
+          <div className="w-6" aria-hidden="true" />
         )) as IconFunctionComponent)
       : null);
